deployment: Add create command (#36)
Adds the deployment create command which enables the creation of
multi-resource deployments from a file definition in one API call. The
main features are two flag overrides (name and version) which allow the
file definition to be used more like a template.

Additionally, a `--request-id` flag is provided which enables setting an
idempotency token for the API calls. By default an auto-generated value
is sent to the API and printed to the Standard Error device so that
users can use the token in case the deployment creation fails.

```
$ dev-cli deployment create -f deployment_example.json --name marcdeploy --version=7.4.1
The deployment creation returned with an error, please use the displayed idempotency token
to recreate the deployment resources
Idempotency token: GMZPMRrcMYqHdmxjIQkHbdjnhPIeBElcwrHwzVlhGUSMXrEIzVXoBykSVRsKncNb
unknown error (status 500)
```

Signed-off-by: Marc Lopez <marc5.12@outlook.com>