go.mod: Update cloud-sdk-go to version v1.0.0-bc4
Updates cloud-sdk-go to version v1.0.0-bc4 in th go.mod file and updates
the required Golang source files. Since a change was merged that is now
able to obtain the underlying response message from the http.Body,
updates all of our tests which relied on the unknown error (status xxx).

Signed-off-by: Marc Lopez <marc5.12@outlook.com>