deployment: Add update command
Adds the `deployment update` command which mainly relies on a JSON file
defnition as the payload to which the specified cluster ID is going to
be updated.

By default it sets prune_orphans=false in the API payload, which causes
the JSON formatted deployment update payload to be treated as a partial
update, meaning, any resources not explicitly specified, will be ignored
and not shut down. To override this behavior toggle `--prune-orphans`
flag.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>