deployment: Fix show resource type command
Fixes the the `deployment show --type=<type>` by removing the harcoded
API calls with the RefId set to `main-<type>`. Additionally solves the
problem of an unspecified (by default) RefId to be autodiscovered with
an API call which obtains the RefId of the resource type when Type is
specified but --ref-id is not.

Additionally adds a `--ref-id` flag which allows the users to override
the default "auto-discovery" behavior and obtain a specific resource if
the RefId is known by the user.

Furthermore, moves all the switch cases to obtain different kind of
resources from the deployment to the more appropriate applicaiton logic
package rather than the `cmd` presentation layer.

Adds an example of this behavior to the help/docs.

Resolves #56.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>