elasticsearch: Add keystore management commands
Adds two new commands, which allow an Elasticsearch cluster's keystore
to be updated and to be shown:

* ecctl deployment elasticsearch keystore set <cluster id> -f def.json
* ecctl deployment elasticsearch keystore show <cluster id>

The specified keystore definition is treated as a partial and the values
in it are used to either create / update the keystore values. Existing
keystore values will not be affected if missing from the request.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>