http: Add api.DefaultTransport in http.Client
Adds the defined sdk `api.DefaultTransport` to the default `http.Client`
used by all the API requests. It fixes the issue where responses with a
StatusCode different to those specified in the swagger spec result in
either an `unknown response` or blank response which makes for very bad
UX. After this change, the body of those requests will be shown to the
user.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>