cmd: Fix init command on unexisting .ecctl folder
There was a small mistake in the directory permission mask when the
folder `~/.ecctl` does not exist, it was previously creating it without
the execution premission which prevents anyone from actually accessing
the folder.

Additionally it prints the location of the configuration that's been
written down at the end of initialising the config.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>