Remove ErrCatchTransport from default http client
Since the newer release of elastic/cloud-sdk-go@v1.0.0-bc7 already adds
the ErrCatchTransport on the construction of the api client, it is no
longer necessary to handle this on the client's side.

Additionally, changes the order on which the application's insecure flag
is set within `ecctl.InitConfig()`. Since the flag insecure comes with a
set default value (false), setting the value before viper.Unmarshal
causes any of the structure's values to be set to the default value
declared in the cobra flags that have been bound to viper.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>