cmd: Move elasticsearch create to deployment API (#67)
This is a breaking change which breaks backwards compatibility with the
previously released behavior of deployment elasticsearch create since
the "--capacity" flag now disappears in favour of a "--size" flag.

Additionally the accepted ElasticsearchPayload from both stdin and the
"--file" flag might be different after this change. In any case there's
a very detailed help and examples section for this command.

Furthermore, specified a version through the "--version" flag is not
required anymore, this is caused by more user-friendly behavior which
now auto-discovers the latest Elastic Stack version for the platform and
uses that as a default.

To keep the ability to track creating deployments a new function called
`depresource.TrackResources` is introduced which has the caveat of not
supporting AppSearch change tracking since the structure is different
under th /deployments API. This is of course less than ideal but it's
also not the scope of that PR to introduce a fucntion which can track
AppSearch workloads. This will be done in a future PR.

Most importantly for this PR is underestanding that the user will now be
faced with a much larger set of configurable options, yet the basic user
won't have too know as much Deployment Templates and toplogies which can
form an Elasticsearch deployment and the its settings, since none of the
flags is required off the bat.

Last, an new function `cmdutil.DecodeDefinition` is introduced so it can
be re-used with commands that accept both a file flag and also a stdin
piping with the same contents as the file.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>