cmd: Migrate kibana create to deployments API
This change mainly moves the kibana create action to use the deployment
API instead of creating the cluster through the cluster APIs.
It makes use of the deployment update API to create the kibana resource
within a deployment which should already exist and have an elasticsearch
resource.

It only provides simple flags and a flat topology with a specified size
for the node memory and the number of zones the deployment resource will
span.

Furthermore, it modifies the depresource.Track parameters to accept an
Orphan key which now allows the tracking of both resources which will be
created / exist and the ones that will be shut down (Orphans).

Signed-off-by: Marc Lopez <marc5.12@outlook.com>