cmd: Migrate apm create to deployments API
This change mainly moves the apm create action to use the deployment
API instead of creating the cluster through the cluster APIs.
It makes use of the deployment update API to create the apm resource
within a deployment which should already exist and have an elasticsearch
resource.

It only provides simple flags and a flat topology with a specified size
for the node memory and the number of zones the deployment resource will
span.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>