cmd: Remove elasticsearch create
The API is introducing a breaking change in the upcoming version of ECE
which will render `ecctl deployment elasticsearch create` not always
usable unless the Deployment template which is being referenced supports
the creation of an Elasticsearch only deployment.

We've discussed that we'll add more options to `ecctl deployment create`
so that it makes the creation of deployments easier, specially for users
which are not acquainted with the API model which is required to create
a single or multi-resource deployment.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>