cmd: Remove newer version check on ecctl version
Removes the check for a newer version against the current ecctl version
since it requires a GitHub Token. It is not the most useful of features
and not well documented. It would be handy to have the ability to check
for newer versions without any request token or authenticaiton. For the
time being I am removing the check.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>