cmd: Fix deployment notes (#114)
Fixes the previously broken deployment note commands.

With the new deployment api and the introduction of the "deployment resource" concept, the api calls to /api/v1/deployments/{deployment_id}/notes no longer work in the same manner.

Until this api endpoint is updated to work using the deployment resource concept, the only deployment IDs it interacts with are the elasticsearch deployments related to the ES IDs (I tested this thoroughly using cURL directly as well as ecctl). So this means we cannot write notes on kibana, apm or appsearch for the time being. I left comments above all related code. In the UI all notes show up on the deployment's main page as well as the ES page.

The UI is using a different method, other than calls to this api endpoint, to leave comments on the deployments. This means that comments made through the UI will not show up on deployment note list.