util: Change default tracking poll interval
Changes the default poll period from 1s to 10s, meaning a 10x increase
to a slower polling period. Additionally adjusts the number of accepted
API failues (404 or timeouts) to only 2 exponential backoffs since now
the polling interval is higher which yields the same effect on ensuring
that the pending plan has finished without polling the API so many
times.

This should reduce memory the binary's CPU and network usage as well as
reduced API backpressure.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>