cmd: improve help wording for ECE specific commands (#169)
Adds "(for ECE installations only)" to the descriptions of the commands which are ECE-specific.

In order to avoid complexity and still provide a good UX, the best course of action is to let the users know which commands will only work in ECE through the --help flag, and the documentation. Users will receive a 401 error if they attempt to run this command.