allocator: Vacate now moves AppSearch instances
Adds support for AppSearch instances to be moved when the `ecctl
platform allocator vacate` is called, adding support for workload
filtering too.

The only difference with other workloads is that plan tracking is not
supported for AppSearch workloads due to elastic/cloud-sdk-go#34.
Instead, an error message is returned after the vacate operations have
been completed which instructrs the user to check the UI to verify that
the vacate has indeed succeeded, of course this is less than ideal, but
once elastic/cloud-sdk-go#34 is resolved, this will no longer be the
case.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>