cmd: init gives you a choice to select default region when ESS is selected (#180)
Whenever a user opts to use ESS in the init flow, they will be prompted to select a default 
region.

```
Select which type of Elastic Cloud offering you will be working with:
  [1] Elasticsearch Service (default).
  [2] Elastic Cloud Enterprise (ECE).
  [3] Elasticsearch Service Private (ESSP).

Please enter your choice: 1

Using "https://api.elastic-cloud.com" as the API endpoint.

Select a region you would like to have as default:
  
  GCP
  [1] us-central1 (Iowa)
  [2] us-east4 (N. Virginia)
  [3] us-west1 (Oregon)
  [4] northamerica-northeast1 (Montreal)
  [5] australia-southeast1 (Sydney)
  [6] europe-west1 (Belgium)
  [7] europe-west2 (London)
  [8] europe-west3 (Frankfurt)
  [9] asia-northeast1 (Tokyo)
  [10] asia-south1 (Mumbai)

  AWS
  [11] us-east-1 (N. Virginia)
  [12] us-west-1 (N. California)
  [13] us-west-2 (Oregon)
  [14] eu-central-1 (Frankfurt)
  [15] eu-west-2 (London)
  [16] eu-west-1 (Ireland)
  [17] ap-northeast-1 (Tokyo)
  [18] ap-southeast-1 (Singapore)
  [19] ap-southeast-2 (Sydney)
  [20] sa-east-1 (São Paulo)

  Azure
  [21] eastus2 (Virginia)
  [22] westus2 (Washington)
  [23] westeurope (Netherlands)
  [24] japaneast (Tokyo)
  [25] southeastasia (Singapore)

Please enter your choice: 
```

Things to note: 

- ESSP defaults to us-west-2 as it's the only available region for the time being
- In order to have some consistency in the list, for the region names, I've decided to go with 
this format instead of the display format when available. This is because GCP does not have a 
display format like the other two.
- The desired thing would be to query an API endpoint and print a list of available regions, said 
endpoint is not available today so in the meantime maintaining a static list of providers is low 
effort and a quick win, any time a new release of ecctl is shipped, new regions can be included 
which makes it a valid distribution mechanism.