vendor: Update elastic/cloud-sdk-go to 8877c73 (#201)
Updates the versioned dependency to the commit `8877c73`. No tag has
been cut since it's beta2 release is upcoming but still some SDK work
needs to be done before that can be published. In any case the bump is
to address a plan tracker bug which caused an erroneous status to be
reported back (Successful plan when in fact the plan had failed).

Signed-off-by: Marc Lopez <marc5.12@outlook.com>