docs: Modify changelog generator for adoc files
Modifies the changelog generator so that it generates an adoc formatted
changelog which is placed in docs/release_notes/${VERSION}.adoc.
Additionally, changes the markdown changelog which is used as the
contents of the GitHub release to only have a reference to the download
links for the binaries and a link to the official docs changelog.
Last, modifies the existing `docs/ecctl-release-notes.asciidoc` file so
the new changelog files are automatically added as part of `make
changelog`.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>