cmd: add instance flags to deployment create command (#224)
Adds instance definition flags to the `ecctl deployment create` command so it is not 
necessary to use a predefined deployment definition JSON file.

Things to note:

- All deployments must at least have one ES and one Kibana instance. This mimics the UI.

- For now each resource type that is being created makes a call to retrieve the deployment 
template definition. This is far from ideal, but the `deployment {kibana | apm | appsearch} 
create` commands use the same functions to update deployments with new instances. 

  Refactoring this falls out of the scope of this PR. I will write a separate issue to integrate the 
`deployment {kibana | apm | appsearch} create` commands into the `deployment update` 
command which will address this issue and have the `deployment update` command act in a 
similar way to the `deployment create` command with the newly introduced flags.

- I have added `params.DeploymentID != "" && len(params.DeploymentID) != 32` to the 
`NewStateless` validation as a temporary measure until the above has been implemented.

- I have removed all output examples from the help text that weren't strictly necessary.

- The links on the help documentation don't work yet, but they will once the ESS public API is released