ecctl: Prefix config parsing error
This change changes the `ecctl.<struct>.Validate()` funcitons
multierrors to use `cloud-sdk-go/pkg/multierror.Prefixed` to improve the
quality of our errors to always use prefixes.

Additionally, changes the behavior when the ecctl.App fails to be
initialized and the config file used is empty (unexisting, or failed to
read) to throw an error which instructs the user to run "ecctl init".

Resolves #223.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>