deployment-template: Adds a new --template-format flag to command (#332)
This patch adds a new `--template-format` flag to the `ecctl platform deployment-template {list | show | pull}` 
commands.

This flag will allow users to specify which information is populated in the response.
If "cluster" is specified, cluster_template is populated in the response. 
If deployment is specified, deployment_template is populated in the response.