deployment: Fix create command against ESS
Fixes the `deployment create` command to be able to pull the deployment
templates by listing them using the API.

Additionally fixes the missing version on statless resources when the
version flag is not specified and it's a flag-based create.

Last, fixes the create tests which weren't testing the create workflow
on the flag-based scenario.

Requires elastic/cloud-sdk-go#178 to be merged.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>