docs: Remove duplicated synopsis from doc files
Copies the markdown generator from github.com/spf13/cobra/doc, modifying
the source code so that it only writes down the `synopsis` header when
the `cmd.Long` is not empty.

Since it's impossible to use any of the existing function signatures to
generate the Markdown docs, copying the source and modifying it locally
is necessary.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>