fix: ignored EC_API_KEY environment variable
Fixes the problem where the `EC_API_KEY` environment variable wasn't
read.

The problem was that `viper` (correctly) unpacked the `EC_API_KEY` as
`api-key`, and since the `mapstructure` tags (in ecctl.Config) reference
`api_key`, the key is correctly un-marshaled when a config file is read,
but not when set via environment variables.

The fix is to set an alias `api_key` to the existing `api-key` key,
which is done in `cmd/root.go:143`.

    v.RegisterAlias("api_key", "api-key")

The rest of the patch is necessary to provide at least some unit tests
to be able to assert this behavior.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>