Update CODEOWNERS (#557)
Much like https://github.com/elastic/cloud-sdk-go/pull/385, There's been a formal agreement to move the ecctl over to the applications team and they are actively working on it and unfortunately it requires Delivery to approve PRs. With this change, Delivery will no longer be a blocker.  This shifts the codeowners to the responsible team. 


@pcsanwald @StaceyKingPoling heads up for awareness.

Co-authored-by: Erik Redding <jeredding@users.noreply.github.com>