Fix goreleaser version (#667)
The release github action is currently broken:

We were using the `latest` gorelease-action, which had changed from 1->2 recently. The main breaking change for us was that `--rm-dist` flag was replaced by `--clean`

But this change causes some other changes:
- Some smaller changes in the goreleaser config
- To run the newest goreleaser, we have to upgrade to golang 1.23.0
- When using go 1.23.0, golangci-lint also needs upgrading
- golanci-lint on the new version also has some config deprecations/changes 

But with all these changes, the local build/lint runs without errors or warnings and the dry-run gorelease also runs through (hopefully the same is the case in the github action).